adlr: hide mouse cursor when typing
show mouse cursor on any mouse event.

This is b/c when the mouse pointer is hidden, the window loses mouse
focus. That's fine when you're typing, but if you want to scroll, you
need that focus to be present. I looked at trying to work around this
by making a bigger diff, but I never got it to work when going in/out
of overview mode, and a bigger diff is a risk when carrying this
downstream.